+++ CHANGELOG.md
@@ -20,3 +20,4 @@
 ## [1.0.4] - 2026-03-25
 
 - Add catalog endpoint
+- Add collection endpoint